refactor: refactor swarm options namespace
