+++ Libplanet.Net.Tests/Consensus/GossipTest.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using Libplanet.Crypto;
 using Libplanet.Net.Consensus;
 using Libplanet.Net.Messages;
+using Libplanet.Net.Options;
 using Libplanet.Net.Transports;
 using Libplanet.Tests.Store;
 using NetMQ;
