+++ Libplanet.Net/Options/BootstrapOptions.cs
@@ -4,7 +4,7 @@ using Libplanet.Net.Messages;
 using Libplanet.Net.Protocols;
 using Libplanet.Net.Transports;
 
-namespace Libplanet.Net
+namespace Libplanet.Net.Options
 {
     public class BootstrapOptions
     {
